requirements.txt
Flask==3.0.3
pandas==2.2.0
scikit-learn==1.4.0
joblib==1.3.2
anthropic==0.25.0
python-dotenv==1.0.1
gunicorn==21.2.0
